www.apply: add helper function for back & next urls
for UpdateJobSeeker process

This is not only a refactoring but also changes the back url from the
process.

Before this change, the user was redirected to the back url from the
apply_session (which was either the SIAE card or the selected job):
now at least the user is redirected to the current process.