Use specific settings for tests
In particular, DEBUG should be false in tests.

Removed NUM_CSRF_SESSION_REQUESTS from tests:
The CSRF was needed because the demo accounts have a form with a CSRF
middleware token. When demo accounts are disabled, there are no
`{% csrf_token %}` in the template, and no need to change the session
(and save it).