Merge JobApplicationQueryset._eligible_job_applications_*
The union causes the database to crawl the table (or indices) three
times. Instead, express the requirements together, and let the database
make informed decisions on how to retrieve the data.

It’s also easier to follow the logic when it’s all in one place.

Performance
===========
Setup: http://localhost:8000/employee_record/list?status=NEW, with a
       SIAE with 15 NEW employee records.
Method: Best of 3

Master: 216 ms
This commit: 175 ms