Allow configuring the adversarial stage start date
Because the adversarial start date lives on the Calendar, all campaigns
are now expected to have a calendar. Closed campaigns won’t need the
start date of the adversarial phase, so not updating those is OK.

The factory has been updated to always attach a calendar to the
evaluation campaign.