Drop Pylint
It is unmaintained for this codebase. Its configuration is not tested,
it is not executed by the CI (likely nor developers), comments to
resolve (or disable) it are missing from many places.

If we want to enable it someday, it’ll take a real effort. In the
meantime, let’s not pay its overhead.