employee_record.models: Reduce `.archivable()` aggressiveness
Currently if an employee record is created for an expired approval it
will be archived at the end of the very same day, it can be too fast
because those particular cases usually involve our support and
asynchronous exchanges.

The 6 months were (arbitrarily) chosen as they seems to be a good
balance between not creating clutter for users doing things on time, but
also not penalizing the other ones, which in turn should help for
support tasks.