Use pytest to override S3_* settings in test
06ba1f4912224518d645c0f565e1280e90484461 prevented overriding the
settings in development. Using an S3 bucket in development is useful,
reallow it.

The previous access to the private `settings._wrapped` now accesses
settings overridden, causing the test to fail (because ALLOWED_HOSTS
cannot be found in the response, which only contains S3_* settings).

The view was re-written to use public methods, at the expense of
triggering warnings (and filtering them out).