Update CreateProlongationForm duration rules
Prolongation can now only be requested 1 year at a time, except for
SENIOR_CDI. That means the special casing of PARTICULAR_DIFFICULTIES in
`get_max_end_at()` can be dropped, that logic being handled in
`CreateProlongationForm`.

Take the cumulative duration for a given reason into account when
computing the maximum end date for a prolongation. Previously, the
MAX_CUMULATIVE_DURATION for that reason was used, resulting in the `max`
attribute of the `end_at` field possibly being beyond the latest
acceptable end date.

test_prolong_approval_view_bad_reason was updated to look for the
error message instead of verifying a random element on the page (that
was removed with this commit).