Reload CreateProlongationForm.reason.help_text on change
Before, the `reason` field was not reloaded with HTMX, so most users did
not see the dynamically defined `help_text` for reason. They could only
see it if there was an error on the form, causing to reload the page.

Instead of reloading the form to toggle the upload panel, always reload
it when a selection is made.

`CheckPrescriberEmailView` and `CheckContactDetailsView` just moved as a
diff artifact. Since the full form is now reloaded when the reason is
changed, it makes more sense to have UpdateFormForReasonView come before
smaller fragments `CheckPrescriberEmailView` and
`CheckContactDetailsView`.