employee_record: Transfer command should failed both the same way
Receiving a record without a line number indicate that we will not be
able to find the associated one and that something is not right on the
other side, so marking it as an error will prevent the file deletion for
further investigation.

If we can't find the record in our database than there is little we can
do, and because SIAE *need* to transfer the record it means that most of
the time there is already a replacement, so no need to block the file
deletion for that kind of error.