Replace dropzone with a file field
Increase security by verifying files uploaded to the app instead of
directly storing them on S3. That improves compatibility with
environments where firewalls block direct access to clever-cloud.

Files are saved using django-storages.

An equivalent of dropzone.js has been implemented to preserve the look
and feel of the application.

File records in the database now use a default date for `last_modified`,
because S3 is no longer the source of truth for the data.

Co-authored-by: hellodeloo <hello@hellodeloo.com>