GEIQ administrative criteria: Add an API code (from LABEL)
We add an api_code field to the GEIQ administrative criteria, which will
help us communicate with LABEL.

We have made the mapping ourselves but not everything is clear yet: some
of the codes don't exactly match to our criteria.

We should discuss quickly, especially about the following:
- "inscrit à Pôle Emploi"
- "demandeur d'emploi depuis moins de 12 mois"

which exist in "les emplois" but nor in LABEL nor in Legifrance.

https://www.legifrance.gouv.fr/jorf/id/JORFTEXT000045398420