[DONOTREVIEW] asp.Commune : Remove the created_by field
It is non-necessary as it had been nitroduced in order to "fakely" add
new Commune and make them appear in the search Commune dropdown.

Not only can we now see the Commune (that was previously invisible) in
the dropdown because its size has been made infinite, but also except
for this case, no Commune has ever been adedd and probably never will.

Also it is misleading because it tends to make us think that it's
possible and OK to add new Commune from the ASP "manually".

And twice so because the manually added ones were most of the time
removed from the queries anyway.

It's for the better to just get rid of those for now.