Rewrite README to encourage virtualenv development
Before this commit, the README was a mix of docker-based deployment, and
virtualenv based deployment. It littered the instructions with
conditional branches, making setup harder than it needs to be.

The Makefile and documented commands had some ambiguity on where they
should run, depending on the setup (USE_VENV=1). Other examples include
file paths (relative to `itou/` when code is available locally, and
directly under `/` in the docker), knowing if PostgreSQL executable
utilities (pg_dump, …) were on the `PATH`.

When a `virtualenv` is not available (because constructed in the
docker), most IDE won’t pickup executable tools and integrate them,
requiring some configuration or duplication between the docker and the
host running them.
The Makefile prerequisites won’t be verified when running e.g. `make
quality` with the docker setup, because the Makefile can’t check the
content of the docker and calls individual commands directly on the
docker.

The team agreed on preferring the `virtualenv` setup, because it
requires less jumping around containers and makes documented commands
easier to read and interpret. There’s no need to copy files between the
host and the docker, resulting in simpler instructions.

The docker setup remains available as an alternative setup, for those
who do not work regularly on the project and would just like to spawn a
working app that executes the project code.

The make recipes to get a shell on a container are preserved, to
keep familiar commands around. They may eventually be moved out of the
makefile to an alias or a helper script.

The backup/restore commands for PostgreSQL are preserved for now,
cleanup is underway with
https://github.com/betagouv/itou-backups/pull/27.

Refs: https://github.com/betagouv/itou/pull/3016#discussion_r1322548291