Add a security.txt for security researchers
Follows RFC 9116 [1]. In case a security vulnerability is identified, a
security.txt helps researchers know how to report the vulnerability.

Expires in a close future, so that the document is reviewed frequently.
The procedure should only take a couple minutes and an out-of-date
document could cause an exploit of the reported vulnerability.

[1] https://www.rfc-editor.org/rfc/rfc9116