approvals : Introduce CancelledApproval model
CancelledApprovals has all the fields that PE needs:
- start_at, end_at
- user details
- number
- SIRET & type of SIAE
- origin of application

all denormalized (to survive eventual SIAE or User deletions), along with the PeNotificationMixin.

Our pole_emploi cron will also read it regularly and send the "latest updates" of the cancelled
approvals to PE.

We at the moment have ~210_000 cancelled approvals for which we have no details in the db;
thanks to Pole Emploi we might recover a lot of these.
Maybe it would be interesting to create lines for the Approvals that we *know* have been cancelled
because of the missing number in the sequence, but the other fields will have to be blank.

We decided as a team that:
- using deleted_at was very tedious, as Django does not use the base managers in related querysets
  so we would have to fish all the uses of `approval__` in any filter, all the uses of `.approval.`
  in the templates...
- that is not even adressing the issues with:
  * cascade-deleting the related objects (linked to the Approval)
  * rewriting all of our UniqueConstraint to take into account the deleted_at field
  * polluting all the filters and methods with taking into account the deleted_at field
  * rewriting our indexes to ignore the "deleted" lines

Another take on it is to have a trigger to save  any deleted line in a huge "deleted stuff"
table that would be append-only and contain the JSON dumped data. This is very much in the works
but what we can't do is to *modify* these lines to for instance "mark" them as sent to PE or not.
We could use some kind of external counter but... Tedious and not very solid.

At the end of the day, the most simple, straight-to-the-point and prgamatic solution is to use
a specialized table to hold exactly the info we need to sent to PE along with the notification
status.

Django docs:
> Base managers aren’t used when querying on related models, or when accessing a one-to-many or
> many-to-many relationship. For example, if the Question model from the tutorial had a deleted
> field and a base manager that filters out instances with deleted=True, a queryset like
> Choice.objects.filter(question__name__startswith='What') would include choices related to
> deleted questions.

Original discussion:
https://itou-inclusion.slack.com/archives/C0412CTV63D/p1695293977257619

Original list of cancelled Approval numbers:
https://www.notion.so/plateforme-inclusion/Extraire-la-liste-des-PASS-IAE-annul-s-51e3cffe2a754db1a1bef401b27e6a35

Holes in the sequence: (mostly, the AI insertion in ~Nov 2021 has had a few issues)
select count(*), FLOOR(number::bigint/1000) from approvals_approval WHERE number not ilike '%XXX%' \
group by FLOOR(number::bigint/1000) order by FLOOR(number::bigint/1000);