employee_record: Add the `asp_measure` to the unique constraint
The same ASP ID can be used in more than one `Convention()`, so the
previous unique constraint was preventing an employer to hire someone in
another of its SIAE when the `asp_id` is the same.

On the ASP side the unicity is `(siret, measure, approval_number)` so
it's probably a good idea to have the same on our side. Properly
replacing `asp_id` by `siret` requires a lot more works, they are more
or less the same but not totally, so for now we are only adding the ASP
measure to fix our current problem.