Refactor SubmitJobApplicationForm
There was a weird separation of concerns between the form and the view.
The form collects three fields, the view prepares the instance by
calling `form.save(commit=False)`, then the view adds a bunch of data to
the object and finally `save()`.

The `ModelForm` shouldn’t `save()` the incomplete object, and does not
fully edit the model. Instead, let the view construct the data and use
the form to validate user-provided input as part of that construction.