Create django-allauth EmailAddress records
The library bases its behavior on the content of the EmailAddress table.
If a user email exists in the EmailAddress table and is verified, then
the user is trying to log in. Otherwise, create an account.

There are currently about 250K EmailAddress records, for about 1M users,
hence about 750K users lack an EmailAddress.
When a user attempts to log in through an identity provider (e.g.
PE Connect in https://itou.sentry.io/issues/2978509211/), the library
searches for a verified EmailAddress record matching that User email.
Because not all users have an EmailAddress record, it will attempt to
create accounts for existing users.

Regression introduced in 32202f6b3fe8b2f7b10efc7d4936ecc2a08689e8, with
upstream change
https://github.com/pennersr/django-allauth/commit/40d05c36dea7a9b88dc6fa5f3b2f2b1b06bfc893.

Before this change, django-allauth looked up the User.email field, but
it now relies on verified EmailAddress records.