geolocate_qs : Allow resolving objects with any geocoding score
We noticed that because of desynchonsization issues, also because the
BAN might be much improved over time, many of the geocoding scores are
not up-to-date anymore.

This options allows us to fix them all, regardless of the current value
and realign the entirety of a queryset for good.

Mostly, for users, low score become high, high scores become very high
(above the reliance threshold of 0.8) and some high scores become low,
but at least we have less surprises when trying to resolve them into
hexa addresses, which is a very slow operation.