__wip : Introduce the "fill_hexa_addresses" script
This script will be useful, after a first update of the user's
coordinates to get the geocoding scores up-to-date, to resolve the
JobSeekerProfile hexa_addresses as massively as we can.

It is a migration step to try and align as best as we can the hexa
addresses on the profile and the user address, in order to someday only
have an address on the JobSeekerProfile.

In the meantime this script would prefill any missing profile address in
bulk, saving time and manual operations.