www.forms: Explicit the expected trigger when specifying `hx-post`
`django-bootstrap5` put the field attributes of the `RadioSelect()`
widget on the wrapping `<div>` instead of the `<input>`, so HTMX was
triggered by the click instead of the input value change.

Specifying `hx-trigger` prevent the `click` event to be captured and we
go back to the previous (and expected) behavior, to avoid any future
mishap it's probably a good idea to set the `hx-trigger` on every fields
with an `hx-post`.