Send prolongation request files through application servers
Verify applications being sent are valid XLSX documents.

Introduce a temporary storage location, to store documents until the
tunnel of a multi-step form is complete. The documents in the temporary
storage expire after 7 days.
Added a test to verify `sync_s3_files` ignores files under that prefix.

A minor user experience regression occurs when the user fills the edit
step, goes to the preview step, then goes back to the edit step. Because
of how browsers handle files, the report file field will be empty when
the user returns to the edit step.
We could retrieve the file handle from the session and preload the file
field, update the widget to offer to delete the previously selected
file, but it seems like the regression is minor enough not to bother.