job_applications: Allow approval number mismatch for employee records
Same rational than the `siret` field:
Ideally we should propose the creation of a new employee record when the
value changes but we are not there yet. Because the `approval_number`
field is updated in `EmployeeRecord._fill_denormalized_fields()` we are
going to allow users to take actions on those employee records so than
can resend it with the new information.

Change of an approval already linked to an employee record is *extremely
rare* (3 cases out of almost 400k) and can only be the result of a human
action in the admin.