asp: Stop using custom exceptions to simplify errors handling
`Commune.DoesNotExist` were rewritten in a custom one which were then
rewritten in `ValidationError`.
Also remove the two static methods on the model because they are only
wrappers for the one on the QuerySet, which can also lead to surprising
effect if we are not paying attention.
This was definitely the case with the `AddressLookupError`, the
exception cannot be raise by `Commune.by_insee_code()`.