www.employee_record: Stop showing "actualisation" message for every ER
After a cloning all employee records with a suspensions or a prolongations
will show the message, but most of the time the SIAE doesn't want (or need) to send the
data because the employee is not an employee anymore. In those cases the
"actualisation" message is getting in the way and the employer can't
clean its dashboard.

Now that we switching from a "do what we tell you" to a "do what you
need" philosophy for employee records, we don't need to show the message
if the employee record actually exists.