employee_record: Remove the notion of "orphan"
Orphans are filtered out by `.for_company()` but when the `siret` field
changes we allow the employer to update and resend the employee record
so it makes little senses to not allow it for the `asp_id` field too.
Especially as it is only used by us as a mean to get a simple but
slightly incorrect unique constraint.