MAP Conseiller entry point
France Travail prescribers will be able to see in their IS (MAP
Conseiller) a list of the job applications they can handle in Les
emplois.
They should see the details of a job application without having to
enter their login credentials.
The first step is to redirect them to Inclusion Connect without
displaying the Login page.