Use Redis for the cache
Previously, the default of `LocMemCache` was used. That cache is
per-process, so on production, each worker has its own cache, which may
lead to surprising behavior. Currently:
- rate limits in DRF were much higher than intended as limits were
  per-process and not global
- the authentication information against PE was cached multiple time,
  making the cache less efficient

For caching non critical data (all the data in the application), a
cache client has been configure to ignore connection errors to the Redis
instance. Errors are reported to Sentry, but the processing goes on, as
if the request key was missing from the cache. That avoids the
application crashing when the cache is missing.

Because the cache is no longer Python, `freeze_time` no longer affects
it. Rewrite `test_get_token_nominal` to account for that.

Stop exposing the REDIS_URL and REDIS_DB settings, they aren’t reused by
the application. The cache is supposed to be accessed through Django
cache connection handler.