Load parent template scripts properly
Most JS scripts are now protected with `htmx.onLoad(() => {...})` to
allow reloading portions of the page with HTMX.
That means most scripts actually require HTMX to be present on the
page.

Some template were adding their scripts without calling to the super
block (which loads HTMX), resulting in the scripts being broken in
various location.

https://itou.sentry.io/issues/4953055957/