companies.models: Better check for `can_use_employee_record`
Conventions are created by importing the data from the source, which is
the ASP's RIAE, if a SIAE company doesn't have a registered convention
then it can't send any employee records.
We can't rely on the middleware since a company created by a staff
member will be considered active, this is mostly done so employers can
hire people as soon as possible while waiting for the legal documents to
be approved and going through everyone.