Add a /check-health endpoint
Allow CleverCloud to monitor the status of the application before to
deploy, should prevent cases where an app that cannot reach the database
is deployed.
https://developers.clever-cloud.com/doc/develop/healthcheck/

CleverCloud probes directly talk to the application through its IP
address, triggering the ALLOWED_HOSTS protection because the probe IP is
in the 10.x.x.x range, and not part of the ALLOWED_HOSTS. In order to
answer, the response must be generated before the Django common
middleware checks for `request.get_host()`.