employee_record.serializers: Use static values for notifications
Fall back to working and recognisable values when serializing a
`EmployeeRecordUpdateNotification()` which is missing some of its
required fields.

When sending a notification the other sides validate everything but will
only use the technical identification fields (`siret`, `mesure`,
`passIae`, `idItou`) to update their approval with the `passDateFin`.
By using those predefined values we will reduce the number of
notifications send back to us with an error because an informations is
missing on our side.