Use HTMX for employers/job search
Avoid reloading the entire page (and scrolling up) when a single filter
changes.

The goal is for the filter form not to be reloaded, to keep the current
scroll position on the page and collapse state. This necessitated
reorganizing the page so that the tabs, heading section and certain form
fields can be reloaded independently with out-of-brand swaps.

Purposely don’t move back to the top of the results. With the redesign
in progress, the filters will move to the top of the page, removing the
need to scroll. Also, as the filter list can be quite tall, users
picking filters from the bottom of the page would be quite annoyed when
the scroll position changes after they picked an element.