Update eligibility diagnosis visibility rules
Hide employers eligibility diagnosis from prescribers, because a
prescriber should be prompted to establish an eligibility diagnosis for
a job seeker, even if an employer has already made one. An employer
diagnosis is only valid for that employer, while an authorized
prescriber diagnosis is valid for all employers, thus more beneficial to
the job seeker.

Previously, the `for_siae` argument to for_job_seeker_and_siae implied
that the viewer was from a company. A prescriber applying for a job
seeker will have `for_siae=company_pk`, although they are not member of
the company and shouldn’t see the employer diagnosis.

Job seekers keep viewing all diagnoses about them.

Employers must now be a member of siae to see the corresponding
diagnosis.