Use a single task per email message to send
Before, when sending any email raised an exception, the whole list of
emails were retried. That could lead to multiple deliveries of the same
email.

Instead, make one task per email, so that the retry only occurs for the
email that failed to send.

The tests were moved a level higher, so instead of verifying the
serialization, the `AsyncEmailBackend` behavior is verified.