Collect sent emails in a database table
When an email is sent, an Email record is created. That will be use to
track the status of sent emails.

Storing the task details in the DB allows to edit emails during the
retry period, e.g. to fix obvious typos.

Extensive logging has been added to help gather knowledge around the
Mailjet API, its documentation lacking in details.

The GIN index has shown to help a lot on a test stable with 3M emails.

Sent emails are cleaned from the database after a year.

In the first version, we’re proxying Mailjet API for email statuses. We
may need to either:
- process the JSON response and display it in HTML, or
- plug the webhook to get more event data

For early investigation purposes, JSON is probably sufficient.
Plugging the webhook would add significant load to the site (about 400K
emails are sent each month). To reduce load, Mailjet throttles the
webhook calls to 1/s, but that’s still a lot of calls.

GEN-1670