Loosen SIAE deletion rules on ASP imports
A user got her antenna deleted twice. In the second deletion, she
received notification emails for an application, but the system deleted
both the antenna and the related job applications.

Let’s prevent deletions when there are job applications, to keep a
record of events happening on the system.
Also, the grace period before an SIAE is deleted has been extended to a
year. That leaves more time for an antenna to become active. There’s
little harm in keeping unused SIAE a while longer, but losing data is
harmful.