UX/UI: Update job applications filters
The new UI has drowdown filters, which do not open a menu when there are
no available filters (e.g. no applications, so no departments). When a
filter is not available, it is hidden from the UI.

Filters are now located in two places:
- quick access from the top bar to the most frequently used filters
- an offcanvas toggled by “Tous les filtres”

There are two forms:
- an HTMX form for the quick access, that reloads the offcanvas form
  along with the results. That keep both forms in sync.
- a regular form, that triggers a page reload, after filters were
  selected from the offcanvas.
In order to keep the state between both forms, hidden fields are added
to each, for the fields they do not show.