Bypass email confirmation page
When users have a link with a confirmation key, the intermediate page
that displays a `<form method="post">` to send that confirmation key.

Yes, a GET request should not have side effects. That’s bad, but the
confirmation prompt is also bad UX. The page is pretty much empty and
has quirks. Since email clients can’t POST, let’s not be too pedantic.