Fix GPS beneficiary creation process
We used the job application process to create beneficiaries for GPS.
This was a quick attempt to make the MVP work but it had a downside:
a company pk is mandatory. The problem is solo prescribers can't provide one.
So they were shown a 500 page.
To be nice with solo prescribers, let's use the PE Singleton only when `gps` is
present in the job application URL.
I know, this piece of code is a real beauty.