SIAE/Job search: simplify handling of departments
The departments were computed by iterating over the list of objects,
forcing to fetch them all to memory. For Paris, with 100 kms radius,
that’s almost 2.5K job offers. That’s a lot of work, where a basic
adjacency list is likely sufficient for users.

Also, the departments field had to be reloaded when the search distance
changed. That would lead to a pretty annoying UI with the upcoming
changes to [Dropdown
filters](https://zeroheight.com/85c89893b/p/207551-buttons): when a
department is selected, the department field is reloaded (with HTMX),
closing the dropdown filter. Hence, a user wanting to select multiple
departments need to open the dropdown, click a department. The results
load, they can then open the dropdown, click a department, and so on.

Instead, use an adjacency list. Users can of course ask for an
impossible combination of filters, such as SIAE 5 kms from Limoges and
in Corrèze. They’ll just get no results, and pick more sensible filters.

The assertNumQueries were outdated, and were updated to better reflect
the actual queries.