UX/UI: Improve company filter on employer search page
There’s only one possible result when the company is used. Clear the
rest of the filters from the URL (except city and distance, which are
used to validate the company value) to fix the following STR:
1. Set distance to 100 kms
2. Filter on a company >5 kms away from the city
3. Set distance to 5 kms

There are no filters active in the UI, yet there are no results, because
the company=PK parameter is present in the URL, but that choice is (now)
invalid and the select2 cannot be initialized with that value.