Run Huey tasks asynchronously in most environments
The review apps have a Redis available, they can use the same settings
as production and demo environments.

Fast machines are made to run scripts, that could possibly send emails.
Make sure the tasks are run immediately.

The test environment remains in immediate mode to facilitate
testing the results of a task inside a test.

The development environment remains in immediate mode, to avoid
requiring developers to spawn a consumer.