Alternative implementation for apply list filters
On the upside, there is only one form, instead of maintaining two forms
in the page, and that behavior is probably what the UX team would love.

On the downside, additional JS is added, making maintenance on this page
more annoying. Small viewports will also spawn a bunch more requests
when in the offcanvas, but users won’t see the results until they close
the offcanvas.

https://itou-inclusion.slack.com/archives/CQ6C3LSAH/p1719415666875599