Refactor tests or job application's acceptance process
Isolate and group form params to ensure default data is correct.
Use the BAN mocked address by default and City PK (instead of a string).

Create a new class specifically for the Accept process.
A nice improvement would be to extract these tests into a new file and
convert them to Pytest. This new class represents a first step in this
direction.