Move JobApplication.get_eligibility_diagnosis to utils
The method needs to evolve to take the viewing user into account.
The views will be responsible for filtering eligibility diagnosis
visible based on the view and user kind.

Create a shared helper to facilitate code reuse.

In the process, circular import issues appeared. The rule to solve them
is to prevent top-level imports from apps that are later in the
dependency chain. Here, Approval depends on eligibility, so eligibility
shouldn’t import from Approval at the top level. Similarly, Approval
depends on user.