Add an AdministrativeCritera.certifiable table column
We have been granted access to the API Particulier
that can certify some administrative criteria.
For the moment, only 3 are in the scope: Allocation Adulte Handicapé,
RSA and Allocation Parent Isolé.

This certification should replace the need for "written proof"
and hence make the "siae evaluation" unnecessary in the (hopefully)
near future.