Extract a User.job_seeker_department helper
Until addresses for job seekers are merged into a common profile, create
a helper that handles how to retrieve the department for a candidate.
The department is often used to limit access to features, and repeating
that logic (or worse, only looking up in one of the two places) is
annoying.

We’ll soon track user department with Matomo, so that logic will have to
be performed for all pages.