Improve pagination.html boost handling
The template hack to detect invalid variables isn’t smart enough to
short-circuit boolean evaluation. Hence, when it verifies an expression
like below, it expects both `boost` and `boost_target` to be defined.
```
boost|default:False or boost_target
```

Instead, simply move the boost check to an outer if block.

Co-authored-by: Xavier Fernandez <xavier.fernandez@beta.gouv.fr>