Add a siae evaluation eligibility column to job app exports
The dates are pretty messy: an campaign evaluates job applications from
the previous year. However, the evaluation campaigns spans over more
than a year, so an application from n-2 can still be under evaluation.

Instead of trying to be smart, always say “eligible” when we’re
considering an auto prescription, regardless of dates.