Remove field SubmitJobApplicationForm.selected_jobs
The data has been validated at step ApplicationJobsView, and stored in
the session. Instead of sending it be to the client and revalidating
input, simply read it back from the session.

Because there were two steps to provide that data, tests would obviously
use both. They have been updated to only use the session.

In order to avoid a useless SQL request to fetch a job description, the
tests aimed to validate the resume file are now spontaneous
applications.