address.models: Make some improvements to `geolocate_qs()`
Use the city as part of the searched, we use a high threshold (>0.8) so
we can add it to help with shared postal code, frequently used street
name or missing diacritical marks.
Without the city, almost all of the results in the 0.6-0.8 range were
good but the outliers were spread in the 0.6-0.7 range. With the city
the result score mostly rise to the 0.65-0.75 range but the good one
get a higher bump than the bad one, if they don't go down as the city
help with disambiguation.

Make possible to configure the reliance score threshold, for job seekers we want
*really* good results as we use them to check ZRR and QPV status, but
for organizations (employers, prescribers, ...) be locate in the correct
city is enough.

Add more verbose informations to better understand what is going on and
what are we doing.