api_particulier: Reuse and close httpx client
The connection was left open after the task was finished, and a new
connection was opened for each request, which incurs significant
overhead.

The class-based design gave the impression that the HTTP client should
be bound to the API client, but client management is better left to the
caller to handle reusing and freeing resources. Use a module instead.

https://www.python-httpx.org/advanced/clients/#why-use-a-client