Split api_particulier.has_required_info
There is a semantic issue: raising `KeyError` for a method that does not
`__getitem__`. Also, it’s surprising to have the _request method fail
because it could not build the parameters for the API.
Let’s check earlier if the job_seeker has enough info for certification.

This change revealed that some test data was not realistic enough:
- in the accept tests, job seekers were missing their birthdate in the
  POST data, which prevented from calling the api particulier. The field
  is present on the page and the birthdate is submitted with POST data.
- the templates were using a job seeker without a birth country, but
  that piece of info is necessary to see the certified badge.