Replace skip button from job seeker signup NIR with an <a>
When implicitly submitting a form (e.g. pressing Enter when in an
<input> element), the form is submitted by the first submit button in
tree order [1].

In this page markup, the first button is the skip button, so we’re
submitting the NIR along with skip=1, instructing the server to skip
this step, not recording the NIR.

We used to redirect the user to the next step when skip=1, might as well
give them a link to that step directly.

[1] https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#implicit-submission