Replace skip button from apply check NIR with an <a>
When implicitly submitting a form (e.g. pressing Enter when in an
<input> element), the form is submitted by the first submit button in
tree order [1].

In this page markup, the first button is the skip button, so we’re
submitting the NIR along with skip=1, instructing the server to skip
this step, not recording the NIR.

We used to redirect the user to the next step when skip=1, might as well
give them a link to that step directly.

In CheckNIRForSenderView, we’re intentionally preventing users to
directly access the search by email feature, to prevent users creating
duplicates. So SearchByEmailForSenderView still requires a
job_seeker_session, which is initialized when the user has entered an
invalid NIR.

[1] https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#implicit-submission