Hide Edit button from job desc page when transferring to self
Transferring applications to one of my companies should go through
internal transfer instead of external transfer. However, when a user
does an external transfer and selects a job description from one of
their companies, they shouldn’t be offered to modify that job
description, but instead to transfer the application.