Shorten employers diagnostics duration to 3 months
Employers must be able to justify administrative criteria they selected
for an IAE hire. The proof is valid for 3 months (we decided 92 days).

Matching the eligibility duration to proof validity period has avoids a
number of edge cases, such has a hire that starts in 5 months with an
RSA criteria: the employer does not have a proof that the candidate will
still be eligible in 2 months.

With administrative criteria being certified, the certification period
now matches the eligibility diagnosis duration.

The `hiring_start_at` remains unchecked, because the subject is pretty
complex. Tracked with
https://www.notion.so/plateforme-inclusion/S-assurer-que-les-dates-d-embauche-en-SIAE-correspondent-une-p-riode-de-validit-du-diagnostic-158e8fa5c35b80959d71c3642d33512e