forms: fix field disabling in readonly forms
In some forms (e.g. `CreateOrUpdateJobSeekerStep1Form`), a field is
disabled if another field is set (`birth_place` and `birth_country` in
the case of the example mentioned). This is triggered by a function in
`utils.js`. The function should not toggle the second field when the
first is disabled, such as in readonly forms.