github: Change when to run the `ci.yml` workflow
The workflow will now trigger when a pull request is associated to the
branch or when a commit is pushed to `master`.

`pull_request` seems to be needed so the workflow is triggered when
clicking the "Approve and run" button for a pull request opened from a
public fork: https://docs.github.com/en/actions/managing-workflow-runs-and-deployments/managing-workflow-runs/approving-workflow-runs-from-public-forks

`merge_group` is needed for the merge queue now that a push on its
`gh-readonly-queue/*` branches doesn't trigger the workflow.