API particulier: Update to breaking changes dateFin=null
In an email on November 15th, 14:23, the following breaking regression
was announced:

> Vous utilisez l'API Revenu de solidarité active du bouquet API Particulier
> et avez prévu d'utiliser les API Allocation de soutien familial et
> Allocation adulte handicapé.
>
> Dans le Swagger et la documentation affichés actuellement sur API
> Particulier, nous indiquons que nous diffusons la date de fermeture du
> droit. Toutefois, cette date n'est pas toujours correcte. En effet, elle
> était calculée par notre API en se basant sur la date de début du droit.
> Cette date de début du droit correspond à la date de première attribution.
> Lorsque le droit est renouvelé, la date de début n'est pas modifiée. Cela
> peut donc renvoyer à une date de fin antérieure à la date à laquelle
> l'appel sur l'API a été effectué avec un Statut "Bénéficiaire".
>
> Pour éviter toute confusion, nous avons fait le choix de passer le champ
> "dateFin" à "null" dans la version actuelle de l'API.

The end date field will always be null in responses.
We were using the end date to set the validity period of a
certification. Since the API only gives “instantaneous” data (is the
person beneficiary now), and the proof must from the past 3 months
before a hire in the law, we can set the validity period to [today,
today + 92 days (~3 months)].